Refactor detectBarcodes function.
